refactor(core): split AI capabilities into typed registry and data files
